+++ core/presentation/build.gradle.kts
@@ -13,5 +13,6 @@ dependencies {
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.lifecycle.viewmodel.android)
     implementation(libs.firebase.messaging.ktx)
+    implementation(libs.bundles.exoplayer)
     testImplementation(libs.junit)
 }
