+++ presentation/intro/src/main/java/com/dhc/intro/start/IntroStartScreen.kt
@@ -1,11 +1,5 @@
 package com.dhc.intro.start
 
-import android.view.View
-import androidx.annotation.DrawableRes
-import androidx.annotation.OptIn
-import androidx.annotation.RawRes
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -14,27 +8,12 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.viewinterop.AndroidView
-import androidx.lifecycle.compose.LifecycleStartEffect
-import androidx.media3.common.MediaItem
-import androidx.media3.common.Player
-import androidx.media3.common.util.UnstableApi
-import androidx.media3.exoplayer.ExoPlayer
-import androidx.media3.ui.AspectRatioFrameLayout
-import androidx.media3.ui.PlayerView
 import com.dhc.designsystem.DhcTheme
 import com.dhc.designsystem.DhcTypoTokens
 import com.dhc.designsystem.button.DhcButton
@@ -43,6 +22,7 @@ import com.dhc.designsystem.button.model.DhcButtonStyle
 import com.dhc.designsystem.title.DhcTitle
 import com.dhc.designsystem.title.DhcTitleState
 import com.dhc.intro.R
+import com.dhc.presentation.component.VideoView
 import com.dhc.presentation.mvi.EventHandler
 
 @Composable
@@ -83,68 +63,6 @@ fun IntroStartScreen(
                 .padding(20.dp)
                 .fillMaxWidth()
                 .align(Alignment.BottomCenter),
-        )
-    }
-}
-
-@OptIn(UnstableApi::class)
-@Composable
-private fun VideoView(
-    @RawRes videoResId: Int,
-    @DrawableRes thumbnailResId: Int,
-    modifier: Modifier = Modifier,
-) {
-    val localContext = LocalContext.current
-    var onPlayWhenReady by remember { mutableStateOf(false) }
-    val exoPlayer = remember {
-        ExoPlayer.Builder(localContext).build().apply {
-            setMediaItem(
-                MediaItem.fromUri("android.resource://${localContext.packageName}/${videoResId}")
-            )
-            repeatMode = Player.REPEAT_MODE_ALL
-            prepare()
-            playWhenReady = true
-        }
-    }
-
-    LifecycleStartEffect(Unit) {
-        val listener = object : Player.Listener {
-            override fun onPlaybackStateChanged(playbackState: Int) {
-                onPlayWhenReady = playbackState == Player.STATE_READY
-            }
-        }
-        exoPlayer.addListener(listener)
-        onStopOrDispose {
-            exoPlayer.release()
-            exoPlayer.removeListener(listener)
-        }
-    }
-
-    Box(modifier = modifier) {
-        AnimatedVisibility(onPlayWhenReady.not()) {
-            Image(
-                painter = painterResource(thumbnailResId),
-                contentScale = ContentScale.Crop,
-                contentDescription = null,
-                modifier = Modifier.fillMaxSize(),
-            )
-        }
-        AndroidView(
-            factory = { context ->
-                PlayerView(context).apply {
-                    useController = false
-                    player = exoPlayer
-                    resizeMode = AspectRatioFrameLayout.RESIZE_MODE_ZOOM
-                }
-            },
-            modifier = Modifier.fillMaxSize(),
-            update = { playerView ->
-                playerView.visibility = if (onPlayWhenReady) {
-                    View.VISIBLE
-                } else {
-                    View.INVISIBLE
-                }
-            }
         )
     }
 }
